pyproject.toml
[tool.poetry]
name = "health-tracker"
version = "0.1.0"
description = ""
authors = ["Dmitrii, <dima12sik@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
pillow = "^10.0.1"
psycopg2-binary = "^2.9.7"
pytils = "^0.4.1"
redis = "^5.0.0"
python-dotenv = "^1.0.0"
django-filter = "^23.3"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.1"
drf-yasg = "^1.21.7"
celery = "^5.3.4"
eventlet = "^0.33.3"
django-celery-beat = "^2.5.0"
django-cors-headers = "^4.2.0"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
